refactor: 도메인 record null 폴백을 fail-fast로 전환하고 캐싱 결함 수정

+++ src/main/java/com/github/yun531/climate/shared/time/TimeShiftUtil.java
@@ -39,7 +39,7 @@ public final class TimeShiftUtil {
     private static int computeShiftHours(LocalDateTime baseHour, LocalDateTime nowHour, int maxShiftHours) {
         long rawHours = ChronoUnit.HOURS.between(baseHour, nowHour);
         if (rawHours <= 0) return 0;
-        return (int) Math.min(rawHours, (long) maxShiftHours);
+        return (int) Math.min(rawHours, maxShiftHours);
     }
 
     private static int computeDayShift(LocalDate baseDate, LocalDate shiftedDate) {
